// DXC_Frontend/src/api/models/hotlineCategoryDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface HotlineCategoryDto {
  publicId?: string
  /** @nullable */
  name: string | null
  /** @nullable */
  description?: string | null
  thuTu?: number
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
  hotlinesCount?: number
}